requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
pylibdmtx @ git+https://github.com/NaturalHistoryMuseum/pylibdmtx.git@00a6e0691db6cd2951d5b4614a9825962f53a7ff
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyzbar==0.1.9
pyzxing==1.0.2
six==1.16.0
